Resolve "Third times a charm: fix spreadsheet loaded types"

Closes #331

See merge request meltano/analytics!350

+++ elt/dbt/models/historical/base/historical_headcount.sql
@@ -6,10 +6,10 @@ WITH source AS (
 ), renamed AS (
 
 
-	SELECT uniquekey as primary_key,
+	SELECT uniquekey::integer as primary_key,
 			month::date as month_of,
 			function as function,
-			employee_cnt as employee_count,
+			employee_cnt::integer as employee_count,
 			
 			--metadata
 			TIMESTAMP 'epoch' + updated_at * INTERVAL '1 second' as updated_at
